# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST de gestion de bibliothèque universitaire.",
        "title": "Bibliothèque API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/emprunts": {
            "get": {
                "description": "Retourne les emprunts actifs pour un utilisateur donne.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunts"
                ],
                "summary": "Lister les emprunts d'un utilisateur",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID utilisateur",
                        "name": "utilisateur_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repositories.EmpruntItem"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Enregistre l'emprunt d'un exemplaire pour un utilisateur.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunts"
                ],
                "summary": "Enregistrer un emprunt",
                "parameters": [
                    {
                        "description": "Demande d'emprunt",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.EmpruntRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/emprunts/retard": {
            "get": {
                "description": "Retourne tous les emprunts en retard.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunts"
                ],
                "summary": "Lister les emprunts en retard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repositories.EmpruntEnRetardItem"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ouvrages/{id}/exemplaires": {
            "get": {
                "description": "Retourne les exemplaires non empruntes pour un ouvrage.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exemplaires"
                ],
                "summary": "Lister les exemplaires disponibles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ouvrage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.ExemplaireResume"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cree un exemplaire pour un ouvrage donne.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exemplaires"
                ],
                "summary": "Creer un exemplaire",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ouvrage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Donnees exemplaire",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreerExemplaireRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ouvrages/{id}/exemplaires/tous": {
            "get": {
                "description": "Retourne tous les exemplaires d'un ouvrage.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exemplaires"
                ],
                "summary": "Lister tous les exemplaires",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ouvrage",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.ExemplaireDetail"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/utilisateurs/rechercher": {
            "get": {
                "description": "Recherche des utilisateurs par nom, prenom, code postal ou numero de telephone.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Rechercher des utilisateurs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nom",
                        "name": "nom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Prenom",
                        "name": "prenom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Code postal",
                        "name": "code_postal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Numero de telephone",
                        "name": "numero_telephone",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repositories.UtilisateurResume"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreerExemplaireRequest": {
            "type": "object",
            "properties": {
                "code_barre": {
                    "type": "string"
                },
                "delai_emprunt_jours": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.EmpruntRequest": {
            "type": "object",
            "required": [
                "code_barre",
                "utilisateur_id"
            ],
            "properties": {
                "code_barre": {
                    "type": "string"
                },
                "utilisateur_id": {
                    "type": "integer"
                }
            }
        },
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "erreur": {
                    "type": "string"
                }
            }
        },
        "controllers.ExemplaireDetail": {
            "type": "object",
            "properties": {
                "code_barre": {
                    "type": "string"
                },
                "date_fin_emprunt": {
                    "type": "string"
                },
                "est_emprunte": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "controllers.ExemplaireResume": {
            "type": "object",
            "properties": {
                "code_barre": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "controllers.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "controllers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "repositories.EmpruntEnRetardItem": {
            "type": "object",
            "properties": {
                "code_barre": {
                    "type": "string"
                },
                "date_fin": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nom": {
                    "type": "string"
                },
                "numero_telephone": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                },
                "titre": {
                    "type": "string"
                }
            }
        },
        "repositories.EmpruntItem": {
            "type": "object",
            "properties": {
                "code_barre": {
                    "type": "string"
                },
                "date_debut": {
                    "type": "string"
                },
                "date_fin": {
                    "type": "string"
                },
                "en_retard": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "titre": {
                    "type": "string"
                }
            }
        },
...
        "repositories.UtilisateurResume": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nom": {
                    "type": "string"
                },
                "numero_telephone": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        }
    }
}